moved navbar component files
