+++ svelte-client/src/lib/components/navbar/navbar.svelte
