+++ svelte-client/src/routes/(app)/+layout.svelte
@@ -1,5 +1,5 @@
 <script>
-    import Navbar from "$lib/navbar.svelte";
+    import Navbar from "$lib/components/navbar/navbar.svelte";
     import { onMount } from 'svelte';
     import { checkAuth } from '$lib';
     import { initNavBar } from "$lib/stores/nav-bar-selected";
